feat(api): :sparkles: add routes for profiles

+++ apps/api/src/services/user.service.ts
@@ -1,4 +1,4 @@
-import { eq } from "drizzle-orm";
+import { eq, and } from "drizzle-orm";
 import { db } from "../database";
 import { user } from "../database/schema";
 
@@ -6,21 +6,35 @@ export const getProfileById = async (id: string) => {
   return await db.select().from(user).where(eq(user.id, id));
 };
 
-export const updateProfileBio = async (id: string, bio: string) => {
-  return await db.update(user).set({ bio }).where(eq(user.id, id));
+export const updateProfileBio = async (
+  id: string,
+  bio: string,
+  userId: string
+) => {
+  if (!userId) throw new Error("Usuario no autenticado.");
+
+  return await db
+    .update(user)
+    .set({ bio })
+    .where(and(eq(user.id, id), eq(user.id, userId)));
 };
 
 export const updateProfile = async (
   id: string,
   fullName: string,
-  email: string
+  email: string,
+  userId: string
 ) => {
+  if (!userId) throw new Error("Usuario no autenticado.");
+
   return await db
     .update(user)
     .set({ name: fullName, email })
-    .where(eq(user.id, id));
+    .where(and(eq(user.id, id), eq(user.id, userId)));
 };
 
-export const deleteProfile = async (id: string) => {
-  return await db.delete(user).where(eq(user.id, id));
+export const deleteProfile = async (id: string, userId: string) => {
+  if (!userId) throw new Error("Usuario no autenticado.");
+
+  return await db.delete(user).where(and(eq(user.id, id), eq(user.id, userId)));
 };
